com.sun.jna.platform.win32.WinDef.WORDByReference holds a WORD (16bit)
WORD is defined to hold 16 bit on windows, so it needs to be accessed 
as short (getShort()).

Closes: #795

+++ contrib/platform/src/com/sun/jna/platform/win32/WinDef.java
@@ -115,7 +115,7 @@ public interface WinDef {
          * @return the value
          */
         public WORD getValue() {
-            return new WORD(getPointer().getInt(0));
+            return new WORD(getPointer().getShort(0));
         }
     }
